requirements.txt
pygame-ce>=2.4.0
numpy>=1.24.0
pillow>=10.0.0
